MakeImage/requirements.txt
PyQt5==5.15.9
Pillow==10.0.0 